progress on #1
Adds `--single` mode to `mapt build`.

Much of it is stubbed in and incomplete, but it sort of works as a
proof-of-concept for nodes/ways.

Two issues:

1) Naive Lua code that uses globals, e.g.:

```lua
function maybe_emit()
  ..
end

function node_function()
  maybe_emit()
end
```

cannot be modularized in the naive manner I was hoping. :( If another
Lua file defines `maybe_emit`, the first file's `node_function` will end
up calling the second one, which is incorrect.

2) Perhaps as a consequence of (1), I'm hitting a tilemaker segfault.

I'm somehow able to process
https://www.openstreetmap.org/relation/13453960 such that tilemaker
thinks it's a linestring.

Naively, this makes sense: it's `type=site`, so it doesn't qualify as a
native multipolygon or a boundary. (I think
https://wiki.openstreetmap.org/wiki/Relation:building may be relevant?)

I think because of (1), my POI code runs in the building phase, calls
Intersects, and the world ends.

I think (2) could be sidestepped if I fixed up my code to enforce
node_keys/way_keys, and to not have the global issue.